+++ frontend/src/main.tsx
@@ -4,14 +4,17 @@ import { createRoot } from "react-dom/client"
 import "./index.css"
 import App from "./App.tsx"
 import { AppProviders } from "@/app/providers.tsx"
+import { AppErrorBoundary } from "@/components/app/AppErrorBoundary.tsx"
 import { ThemeProvider } from "@/components/theme-provider.tsx"
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <ThemeProvider>
-      <AppProviders>
-        <App />
-      </AppProviders>
-    </ThemeProvider>
+    <AppErrorBoundary>
+      <ThemeProvider>
+        <AppProviders>
+          <App />
+        </AppProviders>
+      </ThemeProvider>
+    </AppErrorBoundary>
   </StrictMode>
 )